Refactor nginx process logic
The original intent of this code was to make sure the main Nginx process
was still running. Using variable-assignment with the old-style backtic
evaluations as well as stringing together several superfluous commands
to evaluate this obfuscates whats going on quite a bit.

Switching to pgrep(1) to search for this process, and using it's exit
code, rather than variable assignment clears things up as well as adding
some more meaningful comments and output.