Turn off the Kong Nginx Daemon
The kong nginx daemon is 'on' by default, which causes Kong to run as a
background daemon instead of a foreground process. This is contrary to
how Cloud Foundry expects applications to run. The default CloudFoundry
Nginx Buildpack configuration sets its daemon off for this reason. See
https://github.com/cloudfoundry/nginx-buildpack/blob/master/fixtures/mainline/nginx.conf

With the daemon off, Kong behaves very similarly to the default Nginx
buildpack (running in the foreground, exiting on a crash, and responding
gracefully to platform SIG* signals)